requirements.txt
altgraph==0.17
appdirs==1.4.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
click==7.1.2
configparser==3.5.0
cssselect==1.1.0
discord.py==1.6.0
EasyProcess==0.3
fake-useragent==0.1.11
future==0.18.2
idna==2.10
ixbrlparse==0.2
litereval==0.0.11
lxml==4.6.2
macholib==1.14
numpy==1.19.4
pandas==1.2.0
parse==1.19.0
pefile==2017.9.3
psutil==5.8.0
pycrypto==2.6.1
pyee==8.1.0
pyinstaller-hooks-contrib==2020.11
pyppdf==0.1.2
pyppeteer==0.2.5
pyquery==1.4.3
python-dateutil==2.8.1
pytz==2020.5
PyVirtualDisplay==2.0
requests==2.25.1
requests-html==0.10.0
selenium==3.141.0
six==1.15.0
soupsieve==2.1
spark-parser==1.6.1
splinter==0.14.0
tqdm==4.56.0
urllib3==1.26.2
w3lib==1.22.0
websockets==8.1
word2number==1.1
xvfbwrapper==0.2.9
pyTelegramBotAPI==3.7.7
